% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{check_qq_time_series}
\alias{check_qq_time_series}
\title{Global quality control for time series}
\usage{
check_qq_time_series(radiation_series, lat, lon)
}
\arguments{
\item{radiation_series}{Global radiation series with the columns qq, qc and ser_date (according to ecad database)}

\item{lat}{latitude in WGS84}

\item{lon}{longitude in WGS84}
}
\description{
quality control for global radiation time series.
Data.frame input has the column names according to the ecad database.
}
\author{
Marieke Dirksen
}
