% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wrdc_data.R
\name{get_country_names}
\alias{get_country_names}
\title{Country names}
\usage{

  get_country_names(url_country = "http://wrdc.mgo.rssi.ru/wrdccgi/protect.exe?data_list_full_protected/t6/t6.html")
}
\arguments{
\item{url_country}{link, standard to the european country list}
}
\description{
Get country names from the wrdc website ((\url{http://wrdc.mgo.rssi.ru/wrdccgi/protect.exe?wrdc/wrdc_new.htm}))
}
