% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wrdc_data.R
\name{get_station_data}
\alias{get_station_data}
\title{Get the station measurements}
\usage{
get_station_data(nm, stn, yr)
}
\arguments{
\item{nm}{name of the country according to the html link}

\item{stn}{name of the station according to the html link}

\item{yr}{year of the measurements to be downloaded}
}
\value{
returns a dataframe with the ser_date (yyyy-mm-dd), qq (radiation in J/cm2), qc (quality flag 0,1 or 2) and the columns qcm and qca with -9 values.
The last two quality control columns are to be filled during another check.
}
\description{
Downloads the table with measurements and rewrites it to a long format. Note that not always all the years are avaiable for download.
}
