% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wrdc_data.R
\name{get_station_meta}
\alias{get_station_meta}
\title{Get the metadata from the station}
\usage{
get_station_meta(nm, stn)
}
\arguments{
\item{nm}{name of the country according to the html link}

\item{stn}{name of the station according to the html link}
}
\description{
downloads the metadata of a station, the fist year of measurement is used to extract the metadata from
}
