% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wrdc_data.R
\name{get_station_names}
\alias{get_station_names}
\title{Stations names}
\usage{
get_station_names(nm)
}
\arguments{
\item{nm}{name of the country according to the html link}
}
\description{
Get station names from the countries on the wrdc website
}
