% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wrdc_data.R
\name{get_station_start}
\alias{get_station_start}
\title{Get the start year of the measurements}
\usage{
get_station_start(nm, stn)
}
\arguments{
\item{nm}{name of the country according to the html link}

\item{stn}{name of the station according to the html link}
}
\description{
downloads the start year of the measurement for a station. The format meta-data is returned in ecad format.
}
