% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{global_radiation_quality_check}
\alias{global_radiation_quality_check}
\title{Global radiation daily quality control check}
\usage{
global_radiation_quality_check(timedate, lat, lon, C = 100,
  C_rare = 50, alpha = 1.2, alpha_rare = 1.5)
}
\arguments{
\item{timedate}{observation days of the global radiation in Date/POSIXct format}

\item{lat}{latitude in WGS84}

\item{lon}{longitude in WGS84}

\item{C}{error margin in W/m2}

\item{C_rare}{rare error margin in W/m2}

\item{alpha}{correction factor}

\item{alpha_rare}{rare correction factor}
}
\description{
Checks if the global radiation does not exceed the physical maximum
}
\author{
Marieke Dirksen
}
