% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{solar_constant}
\alias{solar_constant}
\title{Solar constant}
\usage{
solar_constant(yearday, S0 = 1367)
}
\arguments{
\item{yearday}{day of the year in julian days}

\item{S0}{Max irradiance clear sky W/m2}
}
\description{
Solar constant
}
\author{
Marieke Dirksen
}
